+++ influxql/statement_rewriter.go
@@ -26,7 +26,7 @@ func rewriteShowFieldKeysStatement(stmt *ShowFieldKeysStatement) (Statement, err
 			{Expr: &VarRef{Val: "fieldKey"}},
 			{Expr: &VarRef{Val: "fieldType"}},
 		}),
-		Sources:    rewriteSources(stmt.Sources, "_fieldKeys"),
+		Sources:    rewriteSources(stmt.Sources, "_fieldKeys", stmt.Database),
 		Condition:  rewriteSourcesCondition(stmt.Sources, nil),
 		Offset:     stmt.Offset,
 		Limit:      stmt.Limit,
@@ -47,6 +47,7 @@ func rewriteShowMeasurementsStatement(stmt *ShowMeasurementsStatement) (Statemen
 		condition = rewriteSourcesCondition(Sources([]Source{stmt.Source}), stmt.Condition)
 	}
 	return &ShowMeasurementsStatement{
+		Database:   stmt.Database,
 		Condition:  condition,
 		Limit:      stmt.Limit,
 		Offset:     stmt.Offset,
@@ -64,7 +65,7 @@ func rewriteShowSeriesStatement(stmt *ShowSeriesStatement) (Statement, error) {
 		Fields: []*Field{
 			{Expr: &VarRef{Val: "key"}},
 		},
-		Sources:    rewriteSources(stmt.Sources, "_series"),
+		Sources:    rewriteSources(stmt.Sources, "_series", stmt.Database),
 		Condition:  rewriteSourcesCondition(stmt.Sources, stmt.Condition),
 		Offset:     stmt.Offset,
 		Limit:      stmt.Limit,
@@ -140,7 +141,7 @@ func rewriteShowTagKeysStatement(stmt *ShowTagKeysStatement) (Statement, error) 
 		Fields: []*Field{
 			{Expr: &VarRef{Val: "tagKey"}},
 		},
-		Sources:    rewriteSources(stmt.Sources, "_tagKeys"),
+		Sources:    rewriteSources(stmt.Sources, "_tagKeys", stmt.Database),
 		Condition:  rewriteSourcesCondition(stmt.Sources, stmt.Condition),
 		Offset:     stmt.Offset,
 		Limit:      stmt.Limit,
@@ -151,22 +152,29 @@ func rewriteShowTagKeysStatement(stmt *ShowTagKeysStatement) (Statement, error) 
 }
 
 // rewriteSources rewrites sources with previous database and retention policy
-func rewriteSources(sources Sources, measurementName string) Sources {
+func rewriteSources(sources Sources, measurementName, defaultDatabase string) Sources {
 	newSources := Sources{}
 	for _, src := range sources {
 		if src == nil {
 			continue
 		}
 		mm := src.(*Measurement)
+		database := mm.Database
+		if database == "" {
+			database = defaultDatabase
+		}
 		newSources = append(newSources,
 			&Measurement{
-				Database:        mm.Database,
+				Database:        database,
 				RetentionPolicy: mm.RetentionPolicy,
 				Name:            measurementName,
 			})
 	}
 	if len(newSources) <= 0 {
-		return append(newSources, &Measurement{Name: measurementName})
+		return append(newSources, &Measurement{
+			Database: defaultDatabase,
+			Name:     measurementName,
+		})
 	}
 	return newSources
 }